move FTs into individual files
