+++ functional_tests/base.py
@@ -1,0 +1,34 @@
+from django.contrib.staticfiles.testing import StaticLiveServerTestCase
+from selenium import webdriver
+from selenium.common.exceptions import WebDriverException
+from selenium.webdriver.common.keys import Keys
+from unittest import skip
+import time
+
+MAX_WAIT = 10
+
+
+class FunctionalTest(StaticLiveServerTestCase):
+
+    @classmethod
+    def setUpClass(cls):
+        super(FunctionalTest, cls).setUpClass()
+        cls.browser = webdriver.Firefox()
+
+    @classmethod
+    def tearDownClass(cls):
+        cls.browser.quit()
+        super(FunctionalTest, cls).tearDownClass()
+
+    def wait_for_row_in_list_table(self, row_text):
+        start_time = time.time()
+        while True:
+            try:
+                table = table = self.browser.find_element_by_id('list_table')
+                rows = table.find_elements_by_tag_name('tr')
+                self.assertIn(row_text, [row.text for row in rows])
+                return
+            except (AssertionError, WebDriverException) as e:
+                if time.time() - start_time > MAX_WAIT:
+                    raise e
+                time.sleep(0.5)
