+++ functional_tests/test_layout_and_styling.py
@@ -1,0 +1,29 @@
+from .base import FunctionalTest
+from selenium.webdriver.common.keys import Keys
+
+
+class LayoutAndStylingTest(FunctionalTest):
+    def test_layout_and_styling(self):
+        # Edith goes to the home page
+        self.browser.get('%s%s' % (self.live_server_url, '/'))
+        self.browser.set_window_size(1024, 768)
+
+        # She notices the input box is nicely centered
+        inputbox = self.browser.find_element_by_id('new_item')
+        self.assertAlmostEqual(
+            inputbox.location['x'] + inputbox.size['width'] / 2,
+            512,
+            delta=10
+        )
+
+        # She starts a new list and sees the input 
+        # is nicely centered there too
+        inputbox.send_keys('testing')
+        inputbox.send_keys(Keys.ENTER)
+        self.wait_for_row_in_list_table('1: testing')
+        inputbox = self.browser.find_element_by_id('new_item')
+        self.assertAlmostEqual(
+            inputbox.location['x'] + inputbox.size['width'] / 2,
+            512,
+            delta=10
+        )
